Bug 444938 - flux actions causes exceptions when no flux host is being set (Fix for exceptions)

+++ org.eclipse.flux.ui.integration/src/org/eclipse/flux/ui/integration/handlers/SyncDisconnectHandler.java
@@ -40,9 +40,11 @@ public class SyncDisconnectHandler extends AbstractHandler {
 		
 		Repository repository = org.eclipse.flux.core.Activator.getDefault().getRepository();
 
-		for (IProject project : selectedProjects) {
-			if (repository.isConnected(project)) {
-				repository.removeProject(project);
+		if (repository != null) {
+			for (IProject project : selectedProjects) {
+				if (repository.isConnected(project)) {
+					repository.removeProject(project);
+				}
 			}
 		}
 		
@@ -51,13 +53,13 @@ public class SyncDisconnectHandler extends AbstractHandler {
 
 	@Override
 	public void setEnabled(Object evaluationContext) {
-		if (evaluationContext instanceof IEvaluationContext) {
+		Repository repository = org.eclipse.flux.core.Activator.getDefault().getRepository();
+		if (repository != null && evaluationContext instanceof IEvaluationContext) {
 			IEvaluationContext evalContext = (IEvaluationContext) evaluationContext;
 			Object selection = evalContext.getVariable(ISources.ACTIVE_CURRENT_SELECTION_NAME);
 			if (selection instanceof ISelection) {
 				IProject[] selectedProjects = getSelectedProjects((ISelection) selection);
 				
-				Repository repository = org.eclipse.flux.core.Activator.getDefault().getRepository();
 				for (IProject project : selectedProjects) {
 					if (repository.isConnected(project)) {
 						setBaseEnabled(true);
@@ -88,4 +90,5 @@ public class SyncDisconnectHandler extends AbstractHandler {
 		
 		return (IProject[]) selectedProjects.toArray(new IProject[selectedProjects.size()]);
 	}
+
 }
